+++ potluck/prompt.py
@@ -1,7 +1,48 @@
 import logging, threading, cmd, json, pprint 
+try:
+    from prettytable import PrettyTable
+except ImportError:
+    pass
 
 log = logging.getLogger(__package__)
 
+
+def tformat(items, **kwargs):
+    """
+    Format a list of arbitrary dictionary items as
+    a pretty table.
+
+    :param list items: items to include in table
+    :return: table
+    :rtype: str
+    """
+    try:
+        if items is not None and len(items) > 0:
+            
+            # Configure table
+            t = PrettyTable(kwargs.get("field_names", items[0].keys()))
+            for name, value in kwargs.items():
+                setattr(t, name, value)
+
+            # Add items to table
+            for item in items:
+                t.add_row([item.get(_) for _ in t.field_names])
+            
+            return str(t)
+
+    # Default to pprint if prettytable is not available
+    except NameError:
+        log.warning("Tip: `pip install prettytable` to improve this output")
+        return pprint.pformat(items)
+
+def tprint(*args, **kwargs):
+    """
+    Print out a list of arbitrary dictionary items
+    as a pretty table.
+
+    :param list items: items to include in table
+    """
+    print(tformat(*args, **kwargs))
 
 def to_hex(data):
     """
@@ -12,7 +53,7 @@ def to_hex(data):
     :rtype: str
     """
     _ord = ord if isinstance(data, str) else int
-    return " ".join(["%x" % _ord(_) for _ in data])
+    return " ".join(["%.2x" % _ord(_) for _ in data])
 
 
 class Task:
@@ -22,7 +63,7 @@ class Task:
     """
     NOP     = 0
     RESUME  = 1
-    READ    = 2     # { addr, size }
+    CALL    = 2
 
     type    = 0
     args    = None
@@ -55,6 +96,7 @@ class Prompt(cmd.Cmd):
     agent       = None
     hook        = None
     event       = None
+    script      = None
 
     def __init__(self, session, agent, *args, **kwargs):
         super(Prompt, self).__init__(*args, **kwargs)
@@ -99,7 +141,8 @@ class Prompt(cmd.Cmd):
                 func["address"], func["moduleName"], func["name"], ", ".join(args), ret))
             print()
 
-            # Open command prompt
+            # Open command prompt and reset command queue
+            self.cmdqueue = [i for i in self.__class__.script]
             self.event.set()
 
         except (json.decoder.JSONDecodeError, KeyError) as e: