+++ docs/source/changelog.rst
@@ -2,6 +2,18 @@
 Changelog
 =========
 
+0.1.0
+^^^^^
+
+January 19, 2021
+
+- Moved angr to extra dependency [angr]
+- Added pretty tables with extra dependency [pretty]
+- Added run command argument for scripted execution
+- Added CALL task to agent for calling exports while hooked
+- Added search to both process and thread interfaces
+- Added enabling hooks from process interface
+
 0.0.1
 ^^^^^